PINK: move save/load engine functions to separate file
